requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
decouple==0.0.7
dj-stripe==2.8.3
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
stripe==4.2.0
urllib3==2.1.0
whitenoise==6.6.0
